refactor: improve export naming
