+++ packages/astro/src/content/index.ts
@@ -4,5 +4,6 @@ export {
 	astroContentAssetPropagationPlugin,
 	astroContentProdBundlePlugin,
 } from './vite-plugin-content-assets.js';
-export { astroContentServerPlugin } from './vite-plugin-content-server.js';
+export { astroContentImportPlugin } from './vite-plugin-content-imports.js';
+export { attachContentServerListeners } from './server-listeners.js';
 export { astroContentVirtualModPlugin } from './vite-plugin-content-virtual-mod.js';
